[P1.1-P1.12] Complete Phase 1: Mumbai city-in-a-box runs end to end
Fixes the five failing city tests (three were wrong assertions, two were real bugs in
surface units: depression depth was poisoned by a NaN accumulator, and the hexagon
fallback returned a frame with no geometry column).

Runs the real pipeline from the cached open data in 2 min 43 s cold and 19 s warm:
168,606 cells at 30 m, 21,296 road segments over 1,719 km, 10,646 surface units all
inside the 0.5-2 ha cap, and an inferred drain graph of 50,110 nodes and 49,983 edges
carrying 1,765 km of pipe. Depressions explain 89.3 % of the chronic register within
150 m against a 60 % target, and 100 % of nodes reach an outfall.

Also closes two gate holes the run exposed. Ruff never checked services/city because an
unanchored exclusion matched any directory named city (ADR-0009); the 30 findings it hid
are fixed. And lane counts were read from a set, whose order varies with Python's hash
seed, so two builds disagreed on 4 segments (ADR-0008).

+++ services/city/varuna_city/config.py
@@ -113,8 +113,8 @@ def city_grid(config: CityConfig) -> CityGrid:
     bottom = math.floor(min(ys) / res) * res
     top = math.ceil(max(ys) / res) * res
 
-    width = int(round((right - left) / res))
-    height = int(round((top - bottom) / res))
+    width = round((right - left) / res)
+    height = round((top - bottom) / res)
     grid = CityGrid(
         crs=config.crs_string,
         res=res,
